Corrige erros de compilação
