+++ Cpp/movimentos_logicos.h
@@ -1,4 +1,3 @@
-#include "base.h"
 #include <vector>
 
 using namespace std;
